fix: Renamed ProductInfo into ProductInfoPage, created ProductInfo and ProductPurchaseCard
