+++ src/components/ProductInfo/style.css
@@ -1,9 +1,3 @@
-.product-info-container {
-    position: relative;
-    height: calc(100vh - 100px);
-    width: 100%;
-}
-
 .preview-icon {
     position: absolute;
     height: 500px;
@@ -36,24 +30,4 @@
     width: 514px;
     left: 650px;
     top: 372px;
-}
-
-.price-and-buy-button-container {
-    position: absolute;
-    font-size:30px;
-    padding: 10px;
-    margin: 10px;
-    left: 1100px;
-    height: 25vh;
-    width: 25vw;
-    top: 102px;
-    border: 2px solid rgb(231, 231, 231);
-}
-
-.buy-button {
-    position: absolute;
-    font-size: 30px;
-    width: 360px;
-    height: 77px;
-    top: 80px;
 }
