+++ src/components/ProductInfoPage/style.css
@@ -1,0 +1,5 @@
+.product-info-container {
+    position: relative;
+    height: calc(100vh - 100px);
+    width: 100%;
+}
