+++ src/store/slices/Cart/index.ts
@@ -26,4 +26,4 @@ export default cartSlice.reducer;
 
 export const selectCartItems = (state: { cart: ICartState }) => state.cart.items;
 
-export const products: IProduct[] = []
+export const products: IProduct[] = [];
